server/requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
et-xmlfile==1.1.0
Flask==3.0.3
Flask-JSON==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.28.1
numpy==1.26.4
Office365-REST-Python-Client==2.5.10
openpyxl==3.1.4
pandas==2.2.2
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
schema==0.7.7
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
XlsxWriter==3.2.0
